final_render/requirements.txt
Flask==2.0.1
line-bot-sdk==1.19.0
psycopg2==2.9.3
